# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation My API",
        "version": "1.0"
    },
    "paths": {
        "/api/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get List Category Data",
                "description": "enter your description here",
                "operationId": "todo",
                "responses": {
                    "default": {
                        "description": "return array model todo"
                    }
                }
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Store Category",
                "description": "-",
                "operationId": "category/store",
                "requestBody": {
                    "description": "form todo",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "is_publish"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "is_publish": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Detail",
                "description": "-",
                "operationId": "category/show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Enkripsi",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "return model todo"
                    }
                }
            },
            "put": {
                "tags": [
                    "Category"
                ],
                "summary": "Update Category",
                "description": "-",
                "operationId": "category/update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Enkripsi",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "form todo",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "is_publish"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "is_publish": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "Delete todo",
                "description": "-",
                "operationId": "category/delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Enkripsi",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Category",
            "description": "Category"
        }
    ]
}